Deploy webapp with API.
The deployment process now includes creating a CloudFront distribution
backed by an S3 bucket. We then pull the most recent version of the
knowmetools/km-web repository, build a version linked to the API
endpoint we're deploying, and upload it to S3.

By combining the deployment of the API and webapp, we can ensure that we
always have a webapp linking to each deployed version of the API and
vice versa.

Closes #2
Closes #4
Fixes #6