Move API environment outside of module.
The environment for the ECS service serving the API is now provided to
the module directly instead of passing in all the required variables so
that the module could assemble the list of variables. This is cleaner
because those variables were only used to assemble the environment and
were not needed by the rest of the module.