Use webhook to trigger API deploy pipeline.
Instead of having CodePipeline poll the API source repository for
changes, we have now configured a webhook so GitHub triggers the
pipeline whenever new code is pushed.

Because this process is identical to the process used to deploy the web
app, we have refactored the required resources into a module.

Closes #18