Deploy to staging environment before production.
The build pipeline has been refactored such that all changes are first
deployed to a staging environment. If that deployment succeeds, the
changes are then promoted to the production environment. This creates a
platform that is easier to expand, particularly if we want to run more
tests between deploying to staging and production.